apps/rag-api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via uvicorn
distro==1.9.0
    # via openai
docx2txt==0.8
    # via -r requirements.in
et-xmlfile==2.0.0
    # via openpyxl
exceptiongroup==1.3.0
    # via anyio
fastapi==0.115.6
    # via -r requirements.in
grpcio==1.75.1
    # via
    #   grpcio-tools
    #   qdrant-client
grpcio-tools==1.75.1
    # via qdrant-client
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx[http2]==0.27.2
    # via
    #   -r requirements.in
    #   openai
    #   qdrant-client
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.11.0
    # via openai
loguru==0.7.2
    # via -r requirements.in
lxml==6.0.2
    # via python-pptx
numpy==2.2.6
    # via
    #   opencv-python-headless
    #   pandas
    #   qdrant-client
openai==1.51.0
    # via -r requirements.in
opencv-python-headless==4.10.0.84
    # via -r requirements.in
openpyxl==3.1.5
    # via -r requirements.in
packaging==25.0
    # via pytesseract
pandas==2.2.3
    # via -r requirements.in
pillow==10.4.0
    # via
    #   -r requirements.in
    #   pytesseract
    #   python-pptx
portalocker==2.10.1
    # via qdrant-client
protobuf==6.32.1
    # via grpcio-tools
pydantic==2.12.0
    # via
    #   fastapi
    #   openai
    #   pydantic-settings
    #   qdrant-client
pydantic-core==2.41.1
    # via pydantic
pydantic-settings==2.4.0
    # via -r requirements.in
pypdf==4.3.1
    # via -r requirements.in
pytesseract==0.3.10
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.9
    # via -r requirements.in
python-pptx==0.6.23
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via uvicorn
qdrant-client==1.9.2
    # via -r requirements.in
regex==2025.9.18
    # via tiktoken
requests==2.32.5
    # via tiktoken
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
starlette==0.41.3
    # via fastapi
tiktoken==0.7.0
    # via -r requirements.in
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   grpcio
    #   openai
    #   pydantic
    #   pydantic-core
    #   pypdf
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   qdrant-client
    #   requests
uvicorn[standard]==0.30.6
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
xlsxwriter==3.2.9
    # via python-pptx

# The following packages are considered to be unsafe in a requirements file:
# setuptools
